Finished the extension (added soil quality to the model)
Soil quality of each patch ranges between 0-1, if the soil quality of a daisies patch is lower than a certain threshold, the daisy dies (and daisies can't be born in patches with lower soil quality than their threshold)

When patches are first made, they are given random soil values (with a possible minimum value)

Daisies take away from a soils threshold (or adds to it). And soil threshold recovers over time.

Also updated the output so it outputs different statistics depending on what you choose to be displayed (in WorldConstants class)